+++ PekisWidget/PekisWidgetBundle.swift
@@ -1,0 +1,9 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct PekisWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        PekisCountdownWidget()
+    }
+}